+++ urdfpy/utils.py
@@ -28,7 +28,7 @@ def rpy_to_matrix(coords):
     R : (3,3) float
         The corresponding homogenous 3x3 rotation matrix.
     """
-    coords = np.asanyarray(coords)
+    coords = np.asanyarray(coords, dtype=np.float64)
     c3, c2, c1 = np.cos(coords)
     s3, s2, s1 = np.sin(coords)
 
@@ -36,7 +36,7 @@ def rpy_to_matrix(coords):
         [c1 * c2, (c1 * s2 * s3) - (c3 * s1), (s1 * s3) + (c1 * c3 * s2)],
         [c2 * s1, (c1 * c3) + (s1 * s2 * s3), (c3 * s1 * s2) - (c1 * s3)],
         [-s2, c2 * s3, c2 * c3]
-    ])
+    ], dtype=np.float64)
 
 
 def matrix_to_rpy(R, solution=1):
@@ -66,7 +66,7 @@ def matrix_to_rpy(R, solution=1):
     coords : (3,) float
         The roll-pitch-yaw coordinates in order (x-rot, y-rot, z-rot).
     """
-    R = np.asanyarray(R)
+    R = np.asanyarray(R, dtype=np.float64)
     r = 0.0
     p = 0.0
     y = 0.0
@@ -87,7 +87,7 @@ def matrix_to_rpy(R, solution=1):
         r = np.arctan2(R[2,1] / np.cos(p), R[2,2] / np.cos(p))
         y = np.arctan2(R[1,0] / np.cos(p), R[0,0] / np.cos(p))
 
-    return np.array([r, p, y])
+    return np.array([r, p, y], dtype=np.float64)
 
 
 def matrix_to_xyz_rpy(matrix):
@@ -121,7 +121,7 @@ def xyz_rpy_to_matrix(xyz_rpy):
     matrix : (4,4) float
         The homogenous transform matrix.
     """
-    matrix = np.eye(4)
+    matrix = np.eye(4, dtype=np.float64)
     matrix[:3,3] = xyz_rpy[:3]
     matrix[:3,:3] = rpy_to_matrix(xyz_rpy[3:])
     return matrix
@@ -144,7 +144,7 @@ def parse_origin(node):
         ``origin`` child. Defaults to the identity matrix if no ``origin``
         child was found.
     """
-    matrix = np.eye(4)
+    matrix = np.eye(4, dtype=np.float64)
     origin_node = node.find('origin')
     if origin_node is not None:
         if 'xyz' in origin_node.attrib:
@@ -259,9 +259,9 @@ def configure_origin(value):
         The created matrix.
     """
     if value is None:
-        value = np.eye(4)
+        value = np.eye(4, dtype=np.float64)
     elif isinstance(value, (list, tuple, np.ndarray)):
-        value = np.asanyarray(value).astype(np.float)
+        value = np.asanyarray(value, dtype=np.float64)
         if value.shape == (6,):
             value = xyz_rpy_to_matrix(value)
         elif value.shape != (4,4):
